слито mpc-hc 4970, 4972 - fixed some warnings. - use size_t for some array indexes. - use int for CString indexes. - Subtitles project: fixed all warnings (x86 and x64), optimization
также попала часть от mpc-hc 4973, 4978.


git-svn-id: svn://mpc-next.ru/mpc-be/trunk@468 d191f3aa-8897-45be-b9a4-56e5fb472b67

+++ src/SubPic/DX9SubPic.cpp
@@ -161,7 +161,7 @@ STDMETHODIMP CDX9SubPic::ClearDirtyRect(DWORD color)
 		/*
 				DWORD* ptr = (DWORD*)bm.bits;
 				DWORD* end = ptr + bm.h*bm.wBytes/4;
-				while(ptr < end) *ptr++ = color;
+				while (ptr < end) *ptr++ = color;
 		*/
 		Unlock(NULL);
 	}
@@ -311,7 +311,7 @@ STDMETHODIMP CDX9SubPic::AlphaBlt(RECT* pSrc, RECT* pDst, SubPicDesc* pTarget)
 
 		D3DCAPS9 d3dcaps9;
 		hr = pD3DDev->GetDeviceCaps(&d3dcaps9);
-		if(d3dcaps9.AlphaCmpCaps & D3DPCMPCAPS_LESS)
+		if (d3dcaps9.AlphaCmpCaps & D3DPCMPCAPS_LESS)
 		{
 			hr = pD3DDev->SetRenderState(D3DRS_ALPHAREF, (DWORD)0x000000FE);
 			hr = pD3DDev->SetRenderState(D3DRS_ALPHATESTENABLE, TRUE);
@@ -367,8 +367,8 @@ CDX9SubPicAllocator::~CDX9SubPicAllocator()
 void CDX9SubPicAllocator::GetStats(int &_nFree, int &_nAlloc)
 {
 	CAutoLock Lock(&ms_SurfaceQueueLock);
-	_nFree = m_FreeSurfaces.GetCount();
-	_nAlloc = m_AllocatedSurfaces.GetCount();
+	_nFree = (int)m_FreeSurfaces.GetCount();
+	_nAlloc = (int)m_AllocatedSurfaces.GetCount();
 }
 
 void CDX9SubPicAllocator::ClearCache()
